Extract associations to separate module
